AshCloak.Info (ash_cloak v0.1.2)

Introspection functions for the AshCloak extension.

Functions

cloak_attributes(dsl_or_extended)

@spec cloak_attributes(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

The attribute or attributes to encrypt. The attribute will be renamed to encrypted_{attribute}, and a calculation with the same name will be added.

cloak_attributes!(dsl_or_extended)

@spec cloak_attributes!(dsl_or_extended :: module() | map()) :: [atom()] | no_return()

The attribute or attributes to encrypt. The attribute will be renamed to encrypted_{attribute}, and a calculation with the same name will be added.

cloak_decrypt_by_default(dsl_or_extended)

@spec cloak_decrypt_by_default(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

A list of attributes that should be decrypted (their calculation should be loaded) by default.

cloak_decrypt_by_default!(dsl_or_extended)

@spec cloak_decrypt_by_default!(dsl_or_extended :: module() | map()) ::
  [atom()] | no_return()

A list of attributes that should be decrypted (their calculation should be loaded) by default.

cloak_on_decrypt(dsl_or_extended)

@spec cloak_on_decrypt(dsl_or_extended :: module() | map()) ::
  {:ok, (any(), any(), any(), any() -> any()) | mfa()} | :error

A function to call when decrypting any value. Takes the resource, field, records, and calculation context. Must return :ok or {:error, error}

cloak_on_decrypt!(dsl_or_extended)

@spec cloak_on_decrypt!(dsl_or_extended :: module() | map()) ::
  ((any(), any(), any(), any() -> any()) | mfa()) | no_return()

A function to call when decrypting any value. Takes the resource, field, records, and calculation context. Must return :ok or {:error, error}

cloak_options(dsl_or_extended)

@spec cloak_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

cloak DSL options

Returns a map containing the and any configured or default values.

cloak_vault(dsl_or_extended)

@spec cloak_vault(dsl_or_extended :: module() | map()) :: {:ok, module()} | :error

The vault to use to encrypt & decrypt the value

cloak_vault!(dsl_or_extended)

@spec cloak_vault!(dsl_or_extended :: module() | map()) :: module() | no_return()

The vault to use to encrypt & decrypt the value